[assets] refactor `_submitClaimRewards` to utils/tx
